requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
python-multipart==0.0.9
pillow==10.2.0
pytesseract==0.3.10
numpy==1.26.3
opencv-python-headless==4.9.0.80